Move createUnitsArray
